refactor(web): migrate jwt-decoder to feature folder
