+++ apps/web/src/messages/en.json
@@ -43,7 +43,6 @@
     "viewOnGithub": "GitHub"
   },
   "Tools": {
-    "jwt-decoder": { "title": "JWT Decoder", "description": "Decode a JWT's header and payload, with a live expiry status." },
     "data-filter-builder": { "title": "Data Filter Builder", "description": "Compose nested filter conditions visually and export them." },
     "object-transformer": { "title": "Object Transformer", "description": "Interactive object transformation playground." },
     "query-builder": { "title": "Query Builder", "description": "Build nested JSONB queries visually and preview live matches." }
@@ -54,13 +53,6 @@
     "copy": "Copy",
     "data": "Data",
     "filter": "Filter",
-    "token": "JWT",
-    "header": "Header",
-    "payload": "Payload",
-    "signature": "Signature",
-    "expiresIn": "expires in {duration}",
-    "expired": "expired",
-    "noExpiry": "no expiry",
     "matched": "{count} matched",
     "notPreviewable": "This operator can't be previewed in the browser",
     "builder": "Builder",
