+++ apps/web/src/messages/zh-TW.json
@@ -43,7 +43,6 @@
     "viewOnGithub": "GitHub"
   },
   "Tools": {
-    "jwt-decoder": { "title": "JWT 解碼器", "description": "解碼 JWT 的 header 與 payload，並顯示即時有效期狀態。" },
     "data-filter-builder": { "title": "資料篩選建構器", "description": "視覺化組合巢狀篩選條件並匯出。" },
     "object-transformer": { "title": "物件轉換器", "description": "互動式物件轉換 playground。" },
     "query-builder": { "title": "查詢建構器", "description": "視覺化建構巢狀 JSONB 查詢，並即時預覽命中結果。" }
@@ -54,13 +53,6 @@
     "copy": "複製",
     "data": "資料",
     "filter": "篩選條件",
-    "token": "JWT",
-    "header": "Header",
-    "payload": "Payload",
-    "signature": "簽章",
-    "expiresIn": "{duration}後過期",
-    "expired": "已過期",
-    "noExpiry": "無有效期",
     "matched": "命中 {count} 筆",
     "notPreviewable": "此 operator 無法在瀏覽器預覽",
     "builder": "Builder",
