+++ apps/web/src/tools/jwt-decoder/index.ts
@@ -1,0 +1,5 @@
+import type { ToolModule } from "@/tools/types";
+
+import { JwtDecoder } from "./ui";
+
+export const tool: ToolModule = { id: "jwt-decoder", Component: JwtDecoder };
