+++ apps/web/src/tools/messages.ts
@@ -2,6 +2,7 @@ import type { LocaleMessages } from "./types";
 
 import { messages as dataFilterTester } from "./data-filter-tester/messages";
 import { messages as jsonbQueryGenerator } from "./jsonb-query-generator/messages";
+import { messages as jwtDecoder } from "./jwt-decoder/messages";
 import { messages as mongoQueryGenerator } from "./mongo-query-generator/messages";
 import { messages as objectFlatten } from "./object-flatten/messages";
 import { messages as typeConverter } from "./type-converter/messages";
@@ -13,4 +14,5 @@ export const toolMessages: LocaleMessages[] = [
   dataFilterTester,
   mongoQueryGenerator,
   jsonbQueryGenerator,
+  jwtDecoder,
 ];
